+++ OpenPlatform/MiniApp/UI/WebAppWebView.swift
@@ -39,58 +39,6 @@ private let tgEventProxySource = "var TelegramWebviewProxyProto = function() {};
     "}; " +
 "var TelegramWebviewProxy = new TelegramWebviewProxyProto();"
 
-
-private let selectionSource = "var css = '*{-webkit-touch-callout:none;} :not(input):not(textarea):not([\"contenteditable\"=\"true\"]){-webkit-user-select:none;}';"
-        + " var head = document.head || document.getElementsByTagName('head')[0];"
-        + " var style = document.createElement('style'); style.type = 'text/css';" +
-        " style.appendChild(document.createTextNode(css)); head.appendChild(style);"
-
-private let videoSource = """
-function disableWebkitEnterFullscreen(videoElement) {
-  if (videoElement && videoElement.webkitEnterFullscreen) {
-    Object.defineProperty(videoElement, 'webkitEnterFullscreen', {
-      value: undefined
-    });
-  }
-}
-
-function disableFullscreenOnExistingVideos() {
-  document.querySelectorAll('video').forEach(disableWebkitEnterFullscreen);
-}
-
-function handleMutations(mutations) {
-  mutations.forEach((mutation) => {
-    if (mutation.addedNodes && mutation.addedNodes.length > 0) {
-      mutation.addedNodes.forEach((newNode) => {
-        if (newNode.tagName === 'VIDEO') {
-          disableWebkitEnterFullscreen(newNode);
-        }
-        if (newNode.querySelectorAll) {
-          newNode.querySelectorAll('video').forEach(disableWebkitEnterFullscreen);
-        }
-      });
-    }
-  });
-}
-
-disableFullscreenOnExistingVideos();
-
-const observer = new MutationObserver(handleMutations);
-
-document.addEventListener('DOMContentLoaded', () => {
-  if (document.body) {
-    observer.observe(document.body, {
-      childList: true,
-      subtree: true
-    });
-  }
-});
-
-function disconnectObserver() {
-  observer.disconnect();
-}
-"""
-
 internal final class WebAppWebView: BaseWebView {
 
     var isTelegramWebApp: Bool = false
@@ -111,6 +59,10 @@ internal final class WebAppWebView: BaseWebView {
         let preferences = WKPreferences()
         preferences.javaScriptEnabled = true
         preferences.javaScriptCanOpenWindowsAutomatically = true
+        if #available(iOS 16.4, *) {
+            // iOS 18+/26 may regress requestFullscreen in WKWebView; prefer legacy video fullscreen path.
+            preferences.isElementFullscreenEnabled = false
+        }
         configuration.preferences = preferences
                 
         if #available(iOS 17.0, *) {
@@ -158,9 +110,25 @@ internal final class WebAppWebView: BaseWebView {
         
         let selectionScript = WKUserScript(source: selectionSource, injectionTime: .atDocumentEnd, forMainFrameOnly: true)
         contentController.addUserScript(selectionScript)
-        
-        let videoScript = WKUserScript(source: videoSource, injectionTime: .atDocumentStart, forMainFrameOnly: false)
-        contentController.addUserScript(videoScript)
+
+        let fullscreenCompatScript = WKUserScript(source: fullscreenCompatSource, injectionTime: .atDocumentStart, forMainFrameOnly: false)
+        contentController.addUserScript(fullscreenCompatScript)
+
+        let iframeFullscreenCompatScript = WKUserScript(source: iframeFullscreenCompatSource, injectionTime: .atDocumentStart, forMainFrameOnly: false)
+        contentController.addUserScript(iframeFullscreenCompatScript)
+
+        #if DEBUG
+        contentController.removeScriptMessageHandler(forName: "miniappxFullscreenDebug")
+        contentController.add(WeakGameScriptMessageHandler { message in
+            if let body = message.body as? [String: Any] {
+                print("[MiniAppXFullscreenDebug] \(body)")
+            } else {
+                print("[MiniAppXFullscreenDebug] \(message.body)")
+            }
+        }, name: "miniappxFullscreenDebug")
+        let fullscreenDebugScript = WKUserScript(source: fullscreenDebugSource, injectionTime: .atDocumentStart, forMainFrameOnly: false)
+        contentController.addUserScript(fullscreenDebugScript)
+        #endif
         
         var handleSchemeScriptMessageImpl: ((WKScriptMessage) -> Void)?
         contentController.add(WeakGameScriptMessageHandler { message in
@@ -229,7 +197,7 @@ internal final class WebAppWebView: BaseWebView {
                 strongSelf.handleSchemeMessage?(message)
             }
         }
-        
+
         self.observeSomthings()
     }